+++ crates/fuel-core/src/graphql_api/worker_service.rs
@@ -96,7 +96,7 @@ mod tests;
 /// The initialization task recovers the state of the GraphQL service database on startup.
 pub struct InitializeTask<TxPool, OnChain, OffChain, ImportResultProvider> {
     chain_id: ChainId,
-    stop_on_error: bool,
+    continue_on_error: bool,
     block_importer: BoxStream<SharedImportResult>,
     tx_pool: TxPool,
     on_chain_database: OnChain,
@@ -111,7 +111,7 @@ pub struct Task<TxPool, D> {
     block_importer: BoxStream<SharedImportResult>,
     database: D,
     chain_id: ChainId,
-    stop_on_error: bool,
+    continue_on_error: bool,
 }
 
 impl<TxPool, D> Task<TxPool, D>
@@ -455,7 +455,7 @@ where
             on_chain_database,
             off_chain_database,
             import_result_provider,
-            stop_on_error,
+            continue_on_error,
         } = self;
 
         let mut task = Task {
@@ -463,7 +463,7 @@ where
             block_importer,
             database: off_chain_database,
             chain_id,
-            stop_on_error,
+            continue_on_error,
         };
 
         while let Some(Some(block)) = task.block_importer.next().now_or_never() {
@@ -517,10 +517,10 @@ where
                     let result = self.process_block(block);
 
                     // In the case of an error, shut down the service to avoid a huge
-                    // de-synchronization between on-chain and off-chain databases..
+                    // de-synchronization between on-chain and off-chain databases.
                     if let Err(e) = result {
                         tracing::error!("Error processing block: {:?}", e);
-                        should_continue = self.stop_on_error;
+                        should_continue = self.continue_on_error;
                     } else {
                         should_continue = true
                     }
@@ -554,7 +554,7 @@ pub fn new_service<TxPool, I, OnChain, OffChain, ImportResultProvider>(
     off_chain_database: OffChain,
     import_result_provider: ImportResultProvider,
     chain_id: ChainId,
-    stop_on_error: bool,
+    continue_on_error: bool,
 ) -> ServiceRunner<InitializeTask<TxPool, OnChain, OffChain, ImportResultProvider>>
 where
     TxPool: ports::worker::TxPool,
@@ -571,6 +571,6 @@ where
         off_chain_database,
         import_result_provider,
         chain_id,
-        stop_on_error,
+        continue_on_error,
     })
 }